+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2023-03-17-1
+
+Changes:
+
+- Simplify
+- Renamed to `flux-authentication-frontend`
+
 ## v2023-03-13-1
 
 Changes: